kt/src/commonTest/kotlin/com/github/ephemient/aoc2021/Day5Test.kt
package com.github.ephemient.aoc2021

import kotlin.test.Test
import kotlin.test.assertEquals

class Day5Test {
    @Test
    fun part1() {
        assertEquals(5, Day5(SAMPLE_INPUT).part1())
    }

    @Test
    fun part2() {
        assertEquals(12, Day5(SAMPLE_INPUT).part2())
    }

    companion object {
        private val SAMPLE_INPUT = listOf(
            "0,9 -> 5,9",
            "8,0 -> 0,8",
            "9,4 -> 3,4",
            "2,2 -> 2,1",
            "7,0 -> 7,4",
            "6,4 -> 2,0",
            "0,9 -> 2,9",
            "3,4 -> 1,4",
            "0,0 -> 8,8",
            "5,5 -> 8,2",
        )
    }
}
